Refactor endpoint functions
