+++ services/endpoint/endpoint.service.js
@@ -2,52 +2,16 @@
 
 const { URL } = require('url')
 const Joi = require('joi')
-const { errorMessages } = require('../dynamic/dynamic-helpers')
 const { BaseJsonService, InvalidParameter } = require('..')
+const { errorMessages } = require('../dynamic-common')
 const { optionalUrl } = require('../validators')
+const { fetchEndpointData, renderEndpointBadge } = require('../endpoint-common')
 
 const blockedDomains = ['github.com', 'shields.io']
 
 const queryParamSchema = Joi.object({
   url: optionalUrl.required(),
 }).required()
-
-const anySchema = Joi.any()
-
-const optionalStringWhenNamedLogoPresent = Joi.alternatives().when(
-  'namedLogo',
-  {
-    is: Joi.string().required(),
-    then: Joi.string(),
-  }
-)
-
-const optionalNumberWhenAnyLogoPresent = Joi.alternatives()
-  .when('namedLogo', { is: Joi.string().required(), then: Joi.number() })
-  .when('logoSvg', { is: Joi.string().required(), then: Joi.number() })
-
-const endpointSchema = Joi.object({
-  schemaVersion: 1,
-  label: Joi.string()
-    .allow('')
-    .required(),
-  message: Joi.string().required(),
-  color: Joi.string(),
-  labelColor: Joi.string(),
-  isError: Joi.boolean().default(false),
-  namedLogo: Joi.string(),
-  logoSvg: Joi.string(),
-  logoColor: optionalStringWhenNamedLogoPresent,
-  logoWidth: optionalNumberWhenAnyLogoPresent,
-  logoPosition: optionalNumberWhenAnyLogoPresent,
-  style: Joi.string(),
-  cacheSeconds: Joi.number()
-    .integer()
-    .min(0),
-})
-  // `namedLogo` or `logoSvg`; not both.
-  .oxor('namedLogo', 'logoSvg')
-  .required()
 
 module.exports = class Endpoint extends BaseJsonService {
   static get category() {
@@ -72,34 +36,8 @@ module.exports = class Endpoint extends BaseJsonService {
     }
   }
 
-  static render({
-    label,
-    message,
-    color,
-    labelColor,
-    namedLogo,
-    logoSvg,
-    logoColor,
-    logoWidth,
-    logoPosition,
-    style,
-    isError,
-    cacheSeconds,
-  }) {
-    return {
-      isError,
-      label,
-      message,
-      color,
-      labelColor,
-      namedLogo,
-      logoSvg,
-      logoColor,
-      logoWidth,
-      logoPosition,
-      style,
-      cacheSeconds,
-    }
+  static render(props) {
+    return renderEndpointBadge(props, { allowLogo: true })
   }
 
   async handle(namedParams, queryParams) {
@@ -116,16 +54,11 @@ module.exports = class Endpoint extends BaseJsonService {
       throw new InvalidParameter({ prettyMessage: 'domain is blocked' })
     }
 
-    const json = await this._requestJson({
-      schema: anySchema,
+    const validated = await fetchEndpointData(this, {
       url,
       errorMessages,
-    })
-    // Override the validation options because we want to reject unknown keys.
-    const validated = this.constructor._validate(json, endpointSchema, {
-      prettyErrorMessage: 'invalid properties',
+      validationPrettyErrorMessage: 'invalid properties',
       includeKeys: true,
-      allowAndStripUnknownKeys: false,
     })
 
     return this.constructor.render(validated)
